tests for config row
