+++ run.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 use Keboola\DbWriter\Exception\ApplicationException;
 use Keboola\DbWriter\Exception\UserException;
 use Keboola\DbWriter\Logger;
-use Keboola\DbWriter\Redshift\Application;
+use Keboola\DbWriter\RedshiftApplication;
 use Monolog\Handler\NullHandler;
 
 define('APP_NAME', 'wr-db-redshift');
@@ -31,7 +31,7 @@ try {
         $logger->setHandlers(array(new NullHandler(Logger::INFO)));
     }
 
-    $app = new Application($config, $logger);
+    $app = new RedshiftApplication($config, $logger);
 
     echo json_encode($app->run());
 } catch (UserException $e) {
