+++ tests/data/functionalRow/in/tables/bad_type.csv
@@ -1,0 +1,10 @@
+"id","name","glasses","created"
+"0","Miro cillik","yes","2017-06-22"
+"1","Ondrej Hlavacek","no","2017-06-22"
+"2","Martin Halamicek","yes","this is not a date"
+"3","Tomas Kacur","yes","2017-06-22"
+"4","Erik Zigo","no","2017-06-22"
+"5","Marc Raiser","sometimes","2017-06-22"
+"6","Petr Simecek","yes","2017-06-22"
+"7","Jakub Matejka","yes","2017-06-22"
+"8","Ondrej Popelka","no","2017-06-22"
