+++ tests/data/functionalRow/in/tables/bad_type.csv.manifest
@@ -1,0 +1,4 @@
+{
+    "id": "simple",
+    "columns": ["id", "name", "glasses", "created"]
+}
