+++ tests/data/functionalRow/in/tables/simple.csv.manifest
@@ -1,0 +1,4 @@
+{
+    "id": "simple",
+    "columns": ["id", "name", "glasses"]
+}
